fixing shooting algorithm

+++ Block.pde
@@ -1,5 +1,5 @@
 class Block extends Entity {
-  Block(int x, int y, int w, int h, ObjectHandler objectHandler, InputHandler inputHandler) {
+  Block(float x, float y, int w, int h, ObjectHandler objectHandler, InputHandler inputHandler) {
     super(x, y, w, h, ObjectID.BLOCK, objectHandler, inputHandler);
   }
